Empirical power analysis
Added generation of simulated data with different effect sizes and Ns, which can then be investigated using the normal statistical methods (here ANOVA).

This allows empirical quantification of the likelihood of getting a statistically signficant effect for a given alpha level given a true effect size and a specified number of participants.